move all page components into routes directory
